Fix pom file generation
This commit fixes the broken commit 54056ef0 where the handling of
square brackets for versions was moved from the pom file into the
CMakeLists.txt of the java subdirectory. A CMake variable expansion was
broken and moreover the created logic also introduced square brackets in
the project version. This was fixed by introducing a new variable
including the square brackets just for upstream dependencies.